requirements.txt
streamlit
langchain-google-genai
langchain-core
tavily-python
python-dotenv
pandas
